requirements.txt
fastapi
uvicorn
python-multipart
httpx
Pillow
pytesseract
openpyxl
notepad requirements.txt
python-dotenv
